+++ packages/components/config/stencil.basic.config.ts
@@ -13,7 +13,7 @@ export const StencilBaseConfig: Config = {
   watchIgnoredRegex: [/\.stories\.(js|jsx|ts|tsx|mdx)$/, /\/stories\//], // ignore storybook files in --watch mode
   enableCache: true,
   tsconfig: 'tsconfig.json',
-  sourceMap: true,
+  hashedFileNameLength: 10,
   autoprefixCss: true,
   plugins: [sass()],
   outputTargets: [