+++ packages/components/src/components/bal-app/bal-app.tsx
@@ -1,19 +1,49 @@
-import { Component, Host, h, Event, EventEmitter } from '@stencil/core'
-import globalScript from '../../global'
+import { Component, Host, h, Event, EventEmitter, Prop, Method } from '@stencil/core'
 import { isBrowser } from '../../utils/browser'
+import { BalMode, initStyleMode } from '../../utils/config'
+import { rIC } from '../../utils/helpers'
+import { Loggable, Logger, LogInstance } from '../../utils/log'
 
 @Component({
   tag: 'bal-app',
 })
-export class App {
+export class App implements Loggable {
+  private focusVisible?: any
+  log!: LogInstance
+
+  @Logger('bal-app')
+  createLogger(log: LogInstance) {
+    this.log = log
+  }
+
+  /**
+   * Mode defines how the styles are loaded. With `css` each component loads his own styles
+   * and with `sass` the component styles needs to be imported with the file `global.components.sass`.
+   */
+  @Prop({ reflect: true }) mode: BalMode = 'css'
+
+  /**
+   * @internal
+   * Tells if the components are ready
+   */
   @Event({ bubbles: true, composed: true }) balAppLoad!: EventEmitter<boolean>
 
   connectedCallback() {
-    globalScript()
+    initStyleMode(this.mode)
   }
 
   componentDidLoad() {
     this.balAppLoad.emit(true)
+    rIC(async () => {
+      import('../../utils/focus-visible').then(module => (this.focusVisible = module.startFocusVisible()))
+    })
+  }
+
+  @Method()
+  async setFocus(elements: HTMLElement[]) {
+    if (this.focusVisible) {
+      this.focusVisible.setFocus(elements)
+    }
   }
 
   render() {