+++ packages/components/src/components/form/bal-input-group/bal-input-group.tsx
@@ -3,6 +3,9 @@ import { BEM } from '../../../utils/bem'
 
 @Component({
   tag: 'bal-input-group',
+  styleUrls: {
+    css: 'bal-input-group.sass',
+  },
 })
 export class InputGroup implements ComponentInterface {
   @Element() el!: HTMLElement